requirements.txt
anthropic==0.40.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
tweepy==4.14.0
praw==7.7.1
python-frontmatter==1.1.0
requests==2.32.3
